Improve NGINX cache configuration
* Have one set of proxy_cache_valid settings

* Express 'expires' with more easily-read "days" format

* Remove redundant proxy_cache_valid directives. There was some
  variation, but not enough for it to be perpetuated.

* Rename shared memory zone from "staticfilecache" to "main_cache"
  because it's not simply a static file cache.

* Add "inactive" parameter to proxy_cache_path, which was defaulting to
  10 minutes.

* Add note about proxy_cache_revalidate (NGINX sends If-Modified-Since
  to origin).  We can't use it because we're still on Ubuntu 12.04LTS,
  with NGINX 1.1.

* Add proxy_cache_lock so concurrent requests don't get sent to origin
  server for the same thing.

* Change some comments.

* Change proxy_cache_use_stale for /thumbp location. Move on to next
  origin server in the event of connection error or timeout via
  proxy_next_upstream, but deliver stale file for a 500 or 503, assuming
  the other server won't respond any better.

* Add headers to individual locations: Cache-Control and
  X-Cache-Status

* Hide X-Powered-By header

* Enable caching of Exhibitions assets for a short time, and lower
  expiration time.

* Enable caching of Primary Source Sets assets

* Make it the cache's default to store 301 responses.

* Remove test.php and phpinfo.php locations within /info location

* Make proxy_cache_key consistent across configuration.

* Remove ability to defeat caching by setting Cache-Bypass header

* Simplify use of proxy_cache_bypass and proxy_no_cache.

The goal of these changes was originally to cache /thumb resources, but
it became apparent that we could do better in caching overall.

It would be great to extend the expiration of frontend assets, but there
are still some that the `frontend` app does not seem to compile and
assign versioned names. The expiration for assets in the frontend
therefore is one minute.

Wordpress assets in wp-includes and in plugins are versioned, but those
in the berkman theme are not. No image assets are versioned. The
expiration for the wordpress site for users who are not logged in (who
do not have the relevant cookies) is one minute for most resources and
one month for the versioned assets.

These changes don't take into account the different requirements for the
development environment. It has uncompiled assets that shouldn't be
cached for so long. The deployments still clear the proxy cache, so this
is OK for the meantime. In the future, we need to conditionally clear
the cache: clear it after a deployment only in the development
environment, and let asset versioning do the job in production.

The _dpla_no_cache cookie for the frontend should still be recognized
because it only applies to admin requests.

The proxy_ignore_headers in the location / block has been modified to
only ignore Cache-Control.  We're following the default NGINX behavior
now of bypassing the cache for responses that have Set-Cookie.

For caching to be optimal, the frontend app needs to be modified to set
the _dpla_no_cache header only when necessary, not with every single
request. It should only be set when an admin logs in or out.

The /saved and /users locations have caching disabled for regular users
who are logged in.