Fix frontend caching, again
Just don't cache the frontend app's HTML. It's too problematic, with
loggged-in users and the app library. Applications should be in charge
of setting good cache headers for authenticated and unauthenticated
requests, and it's bad to second-guess them in the proxy configuration.

With this change, and some work on the frontend app, it will be possible
some day to forego the cache-purging that we usually do after a
deployment.